Allow IMU logs of longer than ~5min (#6501)
* Allow IMU logs of longer than ~5min

* Don't overwrite IMU logs that exist already

+++ engine/robotToEngineImplMessaging.h
@@ -88,7 +88,7 @@ private:
   // These methods actually do the creation of messages and sending
   // (via MessageHandler) to the physical robot
 
-  uint8_t _imuSeqID = 0;
+  uint32_t _imuSeqID = 0;
   std::ofstream _imuLogFileStream;
   TracePrinter _traceHandler;
